+++ internal/otelcollector/config/builder/common/otlpexporter/config_builder.go
@@ -1,4 +1,4 @@
-package otlpoutput
+package otlpexporter
 
 import (
 	"context"
