+++ internal/otelcollector/config/builder/common/otlpexporter/env_vars.go
@@ -1,4 +1,4 @@
-package otlpoutput
+package otlpexporter
 
 import (
 	"context"
